# deploy/demo/openapi/specs/openapi-localhost:9082.yml
consumes:
- application/json
produces:
- application/json
schemes:
- https
swagger: "2.0"
info:
  title: (Aries Agent REST Server) of aries-framework-go.
  license:
    name: 'SPDX-License-Identifier: Apache-2.0'
  version: 0.1.0
host: localhost:9082
paths:
...
  /vdr/did:
    post:
      tags:
      - vdr
      summary: Saves a did document with the friendly name.
      operationId: saveDIDReq
      parameters:
      - description: Params for creating the did document
        name: Params
        in: body
        schema:
          $ref: '#/definitions/CreateDIDRequest'
      responses:
        default:
          $ref: '#/responses/genericError'
  /vdr/did/create:
    post:
      tags:
      - vdr
      summary: Create a did document.
      operationId: createDIDReq
      responses:
        "200":
          $ref: '#/responses/documentRes'
        default:
          $ref: '#/responses/genericError'
  /vdr/did/records:
    get:
      description: Retrieves the did records
      tags:
      - vdr
      operationId: getDIDRecords
      responses:
        "200":
          $ref: '#/responses/didRecordResult'
        default:
          $ref: '#/responses/genericError'
  /vdr/did/resolve/{id}:
    get:
      description: Resolve did
      tags:
      - vdr
      operationId: resolveDIDReq
      parameters:
      - type: string
        x-go-name: ID
        description: DID ID - pass the did
        name: id
        in: path
        required: true
      responses:
        "200":
          $ref: '#/responses/resolveDIDRes'
        default:
          $ref: '#/responses/genericError'
  /vdr/did/{id}:
    get:
      tags:
      - vdr
      summary: Gets did document with the friendly name.
      operationId: getDIDReq
      parameters:
      - type: string
        x-go-name: ID
        description: DID ID - pass the did
        name: id
        in: path
        required: true
      responses:
        "200":
          $ref: '#/responses/documentRes'
        default:
          $ref: '#/responses/genericError'
...
definitions:
...
  Code:
    type: integer
    format: int32
    title: Code is the error code of command errors.
    x-go-package: github.com/hyperledger/aries-framework-go/pkg/controller/command
...
  CreateDIDRequest:
    type: object
    title: CreateDIDRequest is model for create did request.
    properties:
      did:
        type: object
        x-go-name: DID
      method:
        type: string
        x-go-name: Method
      opts:
        type: object
        additionalProperties:
          type: object
        x-go-name: Opts
    x-go-package: github.com/hyperledger/aries-framework-go/pkg/controller/command/vdr
...
  Record:
    type: object
    title: Record model.
    properties:
      id:
        type: string
        x-go-name: ID
      name:
        type: string
        x-go-name: Name
    x-go-package: github.com/hyperledger/aries-framework-go/pkg/store/did
...
  genericErrorBody:
    type: object
    properties:
      code:
        $ref: '#/definitions/Code'
      message:
        type: string
        x-go-name: Message
    x-go-package: github.com/hyperledger/aries-framework-go/pkg/controller/rest
responses:
...
  didRecordResult:
    description: |-
      didRecordResult model

      This is used to return did records.
    schema:
      type: array
      items:
        $ref: '#/definitions/Record'
...
  documentRes:
    description: |-
      documentRes model

      This is used for returning query connection result for single record search
...
  genericError:
    description: genericError is aries rest api error response
    schema:
      $ref: '#/definitions/genericErrorBody'
...
  resolveDIDRes:
    description: |-
      resolveDIDRes model

      This is used for returning DID resolution response.
...